moved firebase config code into new module
